Set up frontend to be run in phoenix
Both in production and dev. Uses a shim for the phoenix dev server
watcher as elm-spa does not listen on stdin for when to exit.
See assets/stdin-shim.sh for more info.

In dev the elm-spa dev server is used and the index page links the user
to the correct url.

The elm app reads a flag to determine the correct url in dev. In
production relative paths will be used.