LiveView and Phoenix 1.7 (#5)
Remove Elm frontend and implement with LiveView.
Elm is super nice, but it complicates the project setup and complexity by needing to know two languages and building the frontend with nodejs. LiveView can give great interactivity and makes the project a quite standard Phoenix project which again makes it easier to contribute to the project. With LiveView nodejs is not needed at all which cuts a lot in complexity.