Undo setting of start_time on match
This is in G5API set when the match starts through Get5 stats plugin.
I will plan for this as well in this app, but might consider moving this
kind of info, stuff that happens after the match started, to another
table and keep this table and model for match config type of info (pre
match start/info needed to to start a martch).