Refactoring database and migrations
Moving away from uuid ids as it makes things a bit more complicated some
places and is strictly not necessary for this app. Matchid in get5 must
be integers (in a string) so uuids can not be used. A second "get5
matchid" is possible but makes stuff like map stats trickier with
foreign keys.

Consolidation some migrations while the project is still largely unused.
Also adding and renaming some fields.